Move AsciiString.Helpers to MemoryExtensions.Byte
